refactor: 최근 뉴스 조회 시 탭에 바로 반영될 수 있도록 커스텀훅 사용
