+++ src/components/layouts/with-side/top-news-halftime/top-news-halftime.tsx
@@ -5,6 +5,7 @@ import TabBar from './tab-bar';
 import { useState } from 'react';
 import TopHalftime from './top-halftime';
 import TopNews from './top-news';
+import useTopNews from '@/lib/hooks/useTopNews';
 
 export interface Tab {
 	label: string;
@@ -13,16 +14,17 @@ export interface Tab {
 
 export default function TopNewsHalftime() {
 	const [selectedIndex, setSelectedIndex] = useState(0);
+	const { news, newsTab } = useTopNews(); // 초기 렌더링 시 조회된 news 타입에 따라 tab 반영
 	const tabs = [
 		{ label: '오늘의 하프타임🔥', isActive: selectedIndex === 0 },
-		{ label: '많이 본 뉴스 TOP5', isActive: selectedIndex === 1 },
+		{ label: newsTab, isActive: selectedIndex === 1 },
 	];
 
 	return (
 		<ComponentFrame className={`p-4 space-y-2.5 ${selectedIndex === 0 ? 'pb-6' : 'pb-0'}`}>
 			<TabBar tabs={tabs} onClick={(i: number) => setSelectedIndex(i)} />
 
-			{selectedIndex === 0 ? <TopHalftime /> : <TopNews />}
+			{selectedIndex === 0 ? <TopHalftime /> : <TopNews news={news} />}
 		</ComponentFrame>
 	);
 }
